firestore_index: Allow making "composite" indexes that only involve 1 field
It's actually possible to have a so-called composite index that nevertheless only involves a single field. This is distinct from single-field index exclusions, which are handled by the google_firestore_field resource. Nevertheless, it was previously impossible to express these in Terraform because we had incorrectly set a min_size == 2 constraint on `fields`. These sorts of indexes are actually required for some use cases.

Add a new example (descending sort on __name__) that demonstrates such a use case.